% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit.R
\name{lit_h}
\alias{lit_h}
\title{LIT correcting for dominance effects}
\usage{
lit_h(y, x, adjustment = NULL, pop_struct = NULL)
}
\arguments{
\item{y}{matrix of traits (n observations by k traits)}

\item{x}{matrix of SNPs (n observations by m SNPs)}

\item{adjustment}{matrix of covariates to adjust traits}

\item{pop_struct}{matrix of PCs that captures population structure}
}
\description{
Internal use for now
}
